Fix more linting failures
