+++ src/kernels/jupyter/interpreter/jupyterCommand.node.ts
@@ -14,7 +14,7 @@ import {
 import { EXTENSION_ROOT_DIR } from '../../../platform/constants.node';
 import { PythonEnvironment } from '../../../platform/pythonEnvironments/info';
 import { JupyterCommands, JupyterDaemonModule } from '../../../webviews/webview-side/common/constants';
-import { IJupyterCommand, IJupyterCommandFactory } from '../types';
+import { IJupyterCommand, IJupyterCommandFactory } from '../types.node';
 
 class InterpreterJupyterCommand implements IJupyterCommand {
     protected interpreterPromise: Promise<PythonEnvironment>;
