+++ src/kernels/jupyter/types.ts
@@ -7,7 +7,6 @@ import { Event } from 'vscode';
 import { SemVer } from 'semver';
 import { Uri, QuickPickItem } from 'vscode';
 import { CancellationToken, Disposable } from 'vscode-jsonrpc';
-import { SpawnOptions, ExecutionResult, ObservableExecutionResult } from '../../platform/common/process/types.node';
 import { IAsyncDisposable, ICell, IDisplayOptions, Resource } from '../../platform/common/types';
 import { JupyterInstallError } from '../../platform/errors/jupyterInstallError';
 import { PythonEnvironment } from '../../platform/pythonEnvironments/info';
@@ -21,7 +20,6 @@ import {
     IJupyterKernelSpec,
     GetServerOptions
 } from '../types';
-import { JupyterCommands } from '../../webviews/webview-side/common/constants';
 
 export type JupyterServerInfo = {
     base_url: string;
@@ -148,80 +146,6 @@ export interface INotebookExporter extends Disposable {
         kernelSpec?: nbformat.IKernelspecMetadata
     ): Promise<nbformat.INotebookContent | undefined>;
     exportToFile(cells: ICell[], file: string, showOpenPrompt?: boolean): Promise<void>;
-}
-
-export interface IJupyterCommand {
-    interpreter(): Promise<PythonEnvironment | undefined>;
-    exec(args: string[], options: SpawnOptions): Promise<ExecutionResult<string>>;
-}
-
-export const IJupyterCommandFactory = Symbol('IJupyterCommandFactory');
-export interface IJupyterCommandFactory {
-    createInterpreterCommand(
-        command: JupyterCommands,
-        moduleName: string,
-        args: string[],
-        interpreter: PythonEnvironment,
-        isActiveInterpreter: boolean
-    ): IJupyterCommand;
-}
-
-export const IJupyterSubCommandExecutionService = Symbol('IJupyterSubCommandExecutionService');
-/**
- * Responsible for execution of jupyter subcommands such as `notebook`, `nbconvert`, etc.
- * The executed code is as follows `python -m jupyter <subcommand>`.
- *
- * @export
- * @interface IJupyterSubCommandExecutionService
- */
-export interface IJupyterSubCommandExecutionService {
-    /**
-     * Checks whether notebook is supported.
-     *
-     * @param {CancellationToken} [cancelToken]
-     * @returns {Promise<boolean>}
-     * @memberof IJupyterSubCommandExecutionService
-     */
-    isNotebookSupported(cancelToken?: CancellationToken): Promise<boolean>;
-    /**
-     * Error message indicating why jupyter notebook isn't supported.
-     *
-     * @returns {Promise<string>}
-     * @memberof IJupyterSubCommandExecutionService
-     */
-    getReasonForJupyterNotebookNotBeingSupported(): Promise<string>;
-    /**
-     * Used to refresh the command finder.
-     *
-     * @returns {Promise<void>}
-     * @memberof IJupyterSubCommandExecutionService
-     */
-    refreshCommands(): Promise<void>;
-    /**
-     * Gets the interpreter to be used for starting of jupyter server.
-     *
-     * @param {CancellationToken} [token]
-     * @returns {(Promise<PythonEnvironment | undefined>)}
-     * @memberof IJupyterInterpreterService
-     */
-    getSelectedInterpreter(token?: CancellationToken): Promise<PythonEnvironment | undefined>;
-    /**
-     * Starts the jupyter notebook server
-     *
-     * @param {string[]} notebookArgs
-     * @param {SpawnOptions} options
-     * @returns {Promise<ObservableExecutionResult<string>>}
-     * @memberof IJupyterSubCommandExecutionService
-     */
-    startNotebook(notebookArgs: string[], options: SpawnOptions): Promise<ObservableExecutionResult<string>>;
-    /**
-     * Gets a list of all locally running jupyter notebook servers.
-     *
-     * @param {CancellationToken} [token]
-     * @returns {(Promise<JupyterServerInfo[] | undefined>)}
-     * @memberof IJupyterSubCommandExecutionService
-     */
-    getRunningJupyterServers(token?: CancellationToken): Promise<JupyterServerInfo[] | undefined>;
 }
 
 export const IJupyterInterpreterDependencyManager = Symbol('IJupyterInterpreterDependencyManager');
