+++ src/platform/export/exportBase.node.ts
@@ -1,7 +1,8 @@
 import { inject, injectable } from 'inversify';
 import * as path from 'path';
 import { CancellationToken, Uri } from 'vscode';
-import { IJupyterSubCommandExecutionService, INotebookImporter } from '../../kernels/jupyter/types';
+import { INotebookImporter } from '../../kernels/jupyter/types';
+import { IJupyterSubCommandExecutionService } from '../../kernels/jupyter/types.node';
 import { IFileSystem } from '../common/platform/types.node';
 import { IPythonExecutionFactory, IPythonExecutionService } from '../common/process/types.node';
